+++ src/promptflow/promptflow/_cli/_pf_azure/_run.py
@@ -62,8 +62,12 @@ Example:
 # Create a run with YAML file:
 pfazure run create -f <yaml-filename>
 # Create a run from flow directory and reference a run:
-pfazure run create --flow <path-to-flow-directory> --data <path-to-data-file> --column-mapping groundtruth='${data.answer}' prediction='${run.outputs.category}' --run <run-name> --variant "${summarize_text_content.variant_0}" --stream  # noqa: E501
-"""
+pfazure run create --flow <path-to-flow-directory> --data <path-to-data-file> --column-mapping groundtruth='${data.answer}' prediction='${run.outputs.category}' --run <run-name> --variant "${summarize_text_content.variant_0}" --stream
+# Create a run from existing workspace flow
+pfazure run create --flow azureml:<flow-name> --data <path-to-data-file> --column-mapping <key-value-pair>
+# Create a run from existing registry flow
+pfazure run create --flow azureml://registries/<registry-name>/models/<flow-name>/versions/<version> --data <path-to-data-file> --column-mapping <key-value-pair>
+"""  # noqa: E501
 
     def add_param_data(parser):
         # cloud pf can also accept remote data