+++ src/promptflow/tests/sdk_cli_azure_test/unittests/test_pf_client.py
@@ -3,12 +3,14 @@
 # ---------------------------------------------------------
 import mock
 import pytest
-from azure.ai.ml.constants._common import AZUREML_RESOURCE_PROVIDER, RESOURCE_ID_FORMAT
 from tests.sdk_cli_azure_test.recording_utilities import is_live
 
 from promptflow import PFClient
 from promptflow._sdk.operations._connection_operations import ConnectionOperations
 from promptflow._sdk.operations._local_azure_connection_operations import LocalAzureConnectionOperations
+
+AZUREML_RESOURCE_PROVIDER = "Microsoft.MachineLearningServices"
+RESOURCE_ID_FORMAT = "/subscriptions/{}/resourceGroups/{}/providers/{}/workspaces/{}"
 
 
 @pytest.mark.sdk_test